Repository exceptions separated.
