+++ src/APP/Infrastructure/adapters/repositories/exceptions/todo/InvalidOperationOnTodo.exception.ts
@@ -1,0 +1,7 @@
+import { InvalidOperation } from '@carbonteq/hexapp';
+
+export class InvalidOperationOnTodo extends InvalidOperation {
+    constructor() {
+      super(`Invalid Operation on Todo`);
+    }
+}
