+++ src/APP/Infrastructure/adapters/repositories/exceptions/todo/TodoAlreadyExists.exception.ts
@@ -1,0 +1,7 @@
+import { AlreadyExistsError } from '@carbonteq/hexapp';
+
+export class TodoAlreadyExists extends AlreadyExistsError {
+    constructor(todoId: string) {
+      super(`Todo with todoId<${todoId}> already exists in DB`);
+    }
+  }
