+++ src/APP/Infrastructure/adapters/repositories/exceptions/todo/TodoNotFound.exception.ts
@@ -1,0 +1,7 @@
+import { NotFoundError } from '@carbonteq/hexapp';
+
+export class TodoNotFound extends NotFoundError {
+  constructor(todoId: string) {
+    super(`Todo with todoId<${todoId}> not found`);
+  }
+}
