+++ src/APP/Infrastructure/adapters/repositories/exceptions/user/InvalidOperationOnUser.ts
@@ -1,0 +1,7 @@
+import { AlreadyExistsError, InvalidOperation, NotFoundError } from '@carbonteq/hexapp';
+
+export class InvalidOperationOnUser extends InvalidOperation {
+    constructor() {
+      super(`Invalid Operation on User`);
+    }
+}
