+++ src/APP/Infrastructure/adapters/repositories/exceptions/user/UserAlreadyExists.exception.ts
@@ -1,0 +1,7 @@
+import { AlreadyExistsError } from "@carbonteq/hexapp";
+
+export class UserAlreadyExists extends AlreadyExistsError {
+    constructor(userId: string) {
+      super(`User with userId<${userId}> already exists in DB`);
+    }
+  }
