+++ src/APP/Infrastructure/adapters/repositories/exceptions/user/UserNotFound.exception.ts
@@ -1,0 +1,7 @@
+import { NotFoundError } from "@carbonteq/hexapp";
+
+export class UserNotFound extends NotFoundError {
+    constructor(userId: string) {
+      super(`User with userId<${userId}> not found`);
+    }
+  }
